BUILD: Wrong link against libldap_r on el6/el7 since package 8.2110.0-1
The issue was introduced in first 8.2110.0 build, when imhttp was
intregrated in our build.

Building imhttp requires libaprutil, which depends on libldap_r,
both devel packages get installed. On el6/el7 we use a custom static
build of libcurl in order to link modules depending on it with a
more recent version than the system default. Configure step of
libcurl was picking up the ldap lib, causing symbol issue when
loading module omelasticsearch. The fix we found for that was to
explicitly disable ldap support on libcurl configure step.

Other issue, the pkgconfig shipped with libaprutil gives -lldap_r
in ldflags. Here it introduces a useless dependency on libldap_r.
To overcome that issue we force APU_LIBS with only -laprutil-1.