BUILD: Fix imhttp unusable in https due to linking issue
It seems we had this issue since first release of 8.2110.0, when imhttp
was added to the build. The module depends on a static build of civetweb
and it has trouble loading the openssl library:

[1662571849] [error] [client ] initialize_openssl: error loading library libcrypto.so
[1662571849] [error] [client ] Error initializing SSL context

The fix used here adds build option -DNO_SSL_DL and instructs the linker
to link the module against openssl libs. The civetweb library won't try
to load the openssl library dynamically at runtime.

We also start cleaning the SPEC file by removing numerous CFLAGS and
LIBS exports known to cause issues or conflicts between make and the
configure script.